Ensure Pelican core tests can be imported during CI
Even when Pelican master is installed first, Poetry sometimes seems to
replace it with the version from PyPI (that does not include tests).
Explicitly uninstalling the PyPI version and *then* installing Pelican
master should resolve this.

Hopefully all of this will be unnecessary as soon as Pelican core tests
can be included in package distributions without problems.